requirements.txt
fastapi
uvicorn
httpx
pydantic
python-docx
pymupdf
openai
tiktoken
pinecone-client
langchain
sentence-transformers
